requirements.txt
streamlit
google-generativeai 
pillow 
altair 
pandas 
python-dotenv